Wrap `useUserStore` in `defineStore`, move it from `src/composables` to `src/stores`, update usages in `UserProfile`, `UsersGrid` and `UsersPage` accordingly
