+++ cest-la-vue/src/views/UsersPage.vue
@@ -1,21 +1,18 @@
 <script setup lang="ts">
 import Loading from "@/components/Loading.vue";
-import { useUserStore } from "@/composables/useUserStore";
+import { useUserStore } from "@/stores/useUserStore";
+import { storeToRefs } from "pinia";
 
-const { users, loadingUsers, errorGettingUsers } = useUserStore();
+const userStore = useUserStore();
+
+// storeToRefs keeps these reactive
+const { errorGettingUsers, loadingUsers } = storeToRefs(userStore);
 </script>
 
 <template>
   <div>
     <div class="fullscreen-message" v-if="errorGettingUsers">Error getting users</div>
     <Loading v-else-if="loadingUsers" />
-    <!-- 
-       Awkward way to pass the same props down to any child route components.
-       This works assuming they all accept the same props, which is a rare use case but fits here
-       https://router.vuejs.org/guide/essentials/passing-props.html#Via-RouterView
-      -->
-    <router-view v-else v-slot="{ Component }">
-      <component :is="Component" :users="users" />
-    </router-view>
+    <router-view v-else />
   </div>
 </template>
